refactoring all chest selectors into generators.
-splitting ChestWorldGen into Surface and Submerged ChestWorldGen.
