+++ Treasure2-1.12.2/src/com/someguyssoftware/treasure2/command/SpawnWitherTreeCommand.java
@@ -5,10 +5,7 @@ package com.someguyssoftware.treasure2.command;
 
 import java.util.Random;
 
-import com.someguyssoftware.gottschcore.positional.Coords;
 import com.someguyssoftware.treasure2.Treasure;
-import com.someguyssoftware.treasure2.config.TreasureConfig;
-import com.someguyssoftware.treasure2.generator.wither.WitherTreeGenerator;
 
 import net.minecraft.command.CommandBase;
 import net.minecraft.command.ICommandSender;
@@ -50,8 +47,9 @@ public class SpawnWitherTreeCommand extends CommandBase {
 
     			Random random = new Random();
     			//BlockPos pos = new BlockPos(x, y, z);
-    			WitherTreeGenerator gen = new WitherTreeGenerator();
-    			gen.generate(world, random, new Coords(x, y, z), TreasureConfig.WITHER_TREE); //Configs.witherTreeConfig);
+    			// TODO get this to work
+//    			WitherTreeGenerator gen = new WitherTreeGenerator();
+//    			gen.generate(world, random, new Coords(x, y, z), TreasureConfig.WITHER_TREE); //Configs.witherTreeConfig);
     		}
 		}
 		catch(Exception e) {
