+++ Treasure2-1.12.2/src/com/someguyssoftware/treasure2/generator/chest/EpicChestGenerator.java
@@ -1,0 +1,34 @@
+/**
+ * 
+ */
+package com.someguyssoftware.treasure2.generator.chest;
+
+import java.util.Random;
+
+import com.someguyssoftware.gottschcore.random.RandomHelper;
+import com.someguyssoftware.treasure2.Treasure;
+import com.someguyssoftware.treasure2.chest.TreasureChestType;
+
+/**
+ * 
+ * @author Mark Gottschling on Dec 4, 2019
+ *
+ */
+public class EpicChestGenerator implements IChestGenerator {
+	
+	/**
+	 * 
+	 */
+	public EpicChestGenerator() {}
+	
+	/**
+	 * Epic will have at least one lock.
+	 */
+	public int randomizedNumberOfLocksByChestType(Random random, TreasureChestType type) {
+		// determine the number of locks to add
+		int numLocks = RandomHelper.randomInt(random, 1, type.getMaxLocks());		
+		Treasure.logger.debug("# of locks to use: {})", numLocks);
+		
+		return numLocks;
+	}
+}
