+++ Treasure2-1.12.2/src/com/someguyssoftware/treasure2/generator/chest/GoldSkullChestGenerator.java
@@ -1,0 +1,78 @@
+/**
+ * 
+ */
+package com.someguyssoftware.treasure2.generator.chest;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Random;
+
+import com.someguyssoftware.gottschcore.loot.LootTable;
+import com.someguyssoftware.gottschcore.random.RandomHelper;
+import com.someguyssoftware.treasure2.Treasure;
+import com.someguyssoftware.treasure2.block.AbstractChestBlock;
+import com.someguyssoftware.treasure2.block.TreasureBlocks;
+import com.someguyssoftware.treasure2.block.TreasureChestBlock;
+import com.someguyssoftware.treasure2.chest.TreasureChestType;
+import com.someguyssoftware.treasure2.enums.Rarity;
+import com.someguyssoftware.treasure2.item.LockItem;
+import com.someguyssoftware.treasure2.item.TreasureItems;
+import com.someguyssoftware.treasure2.loot.TreasureLootTableMaster.SpecialLootTables;
+import com.someguyssoftware.treasure2.tileentity.AbstractTreasureChestTileEntity;
+
+/**
+ * 
+ * @author Mark Gottschling on Dec 4, 2019
+ *
+ */
+public class GoldSkullChestGenerator implements IChestGenerator {
+	
+	/**
+	 * 
+	 */
+	public GoldSkullChestGenerator() {}
+	
+	/*
+	 * @param random
+	 * @param chestRarity
+	 * @return
+	 */
+	@Override
+	public LootTable selectLootTable(Random random, final Rarity chestRarity) {
+		return Treasure.LOOT_TABLES.getSpecialLootTable(SpecialLootTables.GOLD_SKULL_CHEST);
+	}
+	 
+	/**
+	 * Always select a wither chest.
+	 */
+	@Override
+	public TreasureChestBlock  selectChest(final Random random, final Rarity rarity) {
+		TreasureChestBlock chest = (TreasureChestBlock) TreasureBlocks.GOLD_SKULL_CHEST;
+		return chest;
+	}
+	
+	/**
+	 * Skull chest will have at least one lock.
+	 */
+	public int randomizedNumberOfLocksByChestType(Random random, TreasureChestType type) {
+		// determine the number of locks to add
+		int numLocks = RandomHelper.randomInt(random, 1, type.getMaxLocks());		
+		Treasure.logger.debug("# of locks to use: {})", numLocks);
+		
+		return numLocks;
+	}
+	
+	/**
+	 * Select Locks from Uncommon and Scare rarities.
+	 * @param chest
+	 */
+	@Override
+	public void addLocks(Random random, AbstractChestBlock chest, AbstractTreasureChestTileEntity te, Rarity rarity) {
+		// select a rarity locks
+		List<LockItem> locks = new ArrayList<>();
+		locks.addAll(TreasureItems.locks.get(Rarity.SCARCE));
+		locks.addAll(TreasureItems.locks.get(Rarity.RARE));		
+		addLocks(random, chest, te, locks);
+		locks.clear();
+	}	
+}
